Rename and separate tool manifest finder and tool manifest editor
